requirements.txt
numpy
scipy
opencv-python
tifffile
Pillow
imageio[ffmpeg]
astropy
astroquery
scikit-image
PySide6
